requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
matplotlib
tensorboard
tqdm
pyyaml
opencv-python